created a project wrapper layout, moved mobile & webpage wrappers to layout folder
